Vercel edits

+++ src/app/modules/user-pages/components/followers-and-following/followers-and-following.component.ts
@@ -23,6 +23,7 @@ import { widthAnim } from 'src/tools/animations';
 import { Subject, takeUntil } from 'rxjs';
 import { NotificationService } from '../../services/notification.service';
 import { INotification } from '../../models/notifications';
+import { supabase } from 'src/app/modules/controls/image/supabase-data';
 
 @Component({
   selector: 'app-followers-and-following',
@@ -69,7 +70,7 @@ export class FollowersAndFollowingComponent implements OnInit, OnDestroy {
     private cd: ChangeDetectorRef,
     private router: Router,
     public userService: UserService,
-  ) { }
+  ) {}
 
   ngOnInit(): void {
     this.renderer.addClass(document.body, 'hide-overflow');
@@ -96,6 +97,15 @@ export class FollowersAndFollowingComponent implements OnInit, OnDestroy {
       });
   }
 
+  noUserpic = '/assets/images/userpic.png';
+
+  getUserpic(user: IUser) {
+    if (user.avatarUrl) {
+      return supabase.storage.from('userpics').getPublicUrl(user.avatarUrl)
+        .data.publicUrl;
+    } else return this.noUserpic;
+  }
+
   switchObject(object: 'following' | 'followers') {
     if (this.searchMode) this.searchOnOff();
     if (object === 'following') {
@@ -106,31 +116,38 @@ export class FollowersAndFollowingComponent implements OnInit, OnDestroy {
     this.cd.markForCheck();
   }
 
-  //подписка текущего пользователя на людей в списке
-  follow(user: IUser) {
-    this.userService.addFollower(user, this.currentUser?.id);
-    this.updateUser(user);
+  async updateUser(user: IUser) {
+    await this.userService.updateUserInSupabase(user);
+  }
 
-    if (this.userService.getPermission('new-follower', this.user)) {
-      const notify: INotification = this.notifyService.buildNotification(
-        'Новый подписчик',
-        `Кулинар ${
-          this.currentUser.fullName
+  //подписка текущего пользователя на людей в списке
+  async follow(user: IUser) {
+    if (this.currentUser.id > 0) {
+      user = this.userService.addFollower(user, this.currentUser?.id);
+      await this.updateUser(user);
+      if (this.userService.getPermission('new-follower', user)) {
+
+        const notify: INotification = this.notifyService.buildNotification(
+          'Новый подписчик',
+          `Кулинар ${this.currentUser.fullName
             ? this.currentUser.fullName
             : '@' + this.currentUser.username
-        } подписался на тебя`,
-        'info',
-        'user',
-        '/cooks/list/' + this.currentUser.id,
-      );
-      this.notifyService.sendNotification(notify, user).subscribe();
-    }
+          } подписался на тебя`,
+          'info',
+          'user',
+          '/cooks/list/' + this.currentUser.id,
+        );
+        await this.notifyService.sendNotification(notify, user)
+      
+      } }
   }
 
   //отписка текущего пользователя от людей в списке
-  unfollow(user: IUser) {
-    this.userService.removeFollower(user, this.currentUser?.id);
-    this.updateUser(user);
+  async unfollow(user: IUser) {
+    if (this.currentUser.id > 0) {
+      user = this.userService.removeFollower(user, this.currentUser?.id);
+      await this.updateUser(user);
+    }
   }
 
   searchOnOff() {
@@ -138,13 +155,6 @@ export class FollowersAndFollowingComponent implements OnInit, OnDestroy {
     this.followingDisplay = this.following;
     this.followersDisplay = this.followers;
     this.searchQuery = '';
-  }
-
-  updateUser(user: IUser) {
-    this.userService
-      .updateUsers(user)
-      .pipe(takeUntil(this.destroyed$))
-      .subscribe();
   }
 
   //переход по ссылке на человека
@@ -182,6 +192,7 @@ export class FollowersAndFollowingComponent implements OnInit, OnDestroy {
       }
     }
   }
+
   clickBackgroundNotContent(elem: Event) {
     if (elem.target !== elem.currentTarget) return;
     this.closeEmitter.emit(true);
